backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
gunicorn==22.0.0
bcrypt==4.2.0
cryptography==43.0.1